+++ src/routes/currency/[id]/+page.ts
@@ -16,8 +16,18 @@ export const load: PageLoad = async ({ parent, params }) => {
 		.eq('currency', params.id);
 	if (venuesError) console.log(venuesError.message);
 
+	const { data: tokens, error: tokensError } = await supabase
+		.from('tokens')
+		.select()
+		.eq('currency', params.id);
+	if (tokensError) console.log(tokensError.message);
+
+	// How many tokens?
+	const count = tokens?.length ?? null;
+
 	return {
 		currency: currency,
-		venues: venues
+		venues: venues,
+		count: count
 	};
 };